Pull from non-"minimal" images; Work on 1 branch
RockyLinux pushed new "minimal" images to dockerhub which are truly
minimal--they don't even have a package manager such as yum or dnf. The
grep command used to parse the different RockyLinux tags from dockerhub
is updated to exclude any tags that have "minimal" in their names.

The updateBuildPush.sh script was written to operate on all branches of
a repo, but this could cause issues such as merge conflicts between
different branches within the repo or from a fork that has an
open/un-merged PR.